Bugfix: Use the "fileLocality" field from dcache, instead of "currentQos", to get the online-ness of a file.
It looks like qos is sometimes incorrect, or at least, not what I
thought it was, since online files can have
fileLocality=ONLINE_AND_NEARLINE but qos=tape. So we use fileLocality
for the online-ness of the file, but still request qos because it
gives us the target qos if there's an outstanding prestage
request. Thanks to Tom Junk for alerting me to the problem.